+++ otrs/release_notes.md
@@ -1,5 +1,7 @@
 #### What's Improved
 - "Search Ticket" action has multiselect "State" to filter.
 - Added new parameter "Time Span Minutes" in Search Ticket" action.
+- Added support for OTRS "DynamicField" in "Create Ticket" Action
 - Added support for OTRS "DynamicField" in "Update Ticket" Action
 - Added support for OTRS "Article" in "Update Ticket" Action
+- Added support for "Article Sender Type" in "Create Ticket" action.